Refactor Sequence Diagram component location to move out of the separate directory
