+++ system/control-plane/global/components/observability/org.wso2.vick.observability.api/src/main/java/org/wso2/vick/observability/api/siddhi/SiddhiStoreQueryTemplates.java
@@ -75,7 +75,7 @@ public enum SiddhiStoreQueryTemplates {
             "or destinationCell == \"${" + Params.DESTINATION_CELL + "}\")\n" +
             "and (\"${" + Params.DESTINATION_MICROSERVICE + "}\" == \"\" " +
             "or destinationVICKService == \"${" + Params.DESTINATION_MICROSERVICE + "}\") " +
-            "and (sourceCell != destinationCell and sourceVICKService != destinationVICKService)\n" +
+            "and (sourceCell != destinationCell or sourceVICKService != destinationVICKService)\n" +
             "within ${" + Params.QUERY_START_TIME + "}L, ${" + Params.QUERY_END_TIME + "}L\n" +
             "per \"${" + Params.TIME_GRANULARITY + "}\"\n" +
             "select AGG_TIMESTAMP, httpResponseGroup, sum(totalResponseTimeMilliSec) as totalResponseTimeMilliSec, " +
