+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/SignIn.js
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-import AuthUtils from "../utils/common/authUtils";
+import AuthUtils from "../utils/api/authUtils";
 import Avatar from "@material-ui/core/Avatar";
 import Button from "@material-ui/core/Button";
 import Checkbox from "@material-ui/core/Checkbox";
