+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/cells/cell/Details.js
@@ -17,7 +17,7 @@
 import CellDependencyView from "./CellDependencyView";
 import ColorGenerator from "../../common/color/colorGenerator";
 import HealthIndicator from "../../common/HealthIndicator";
-import HttpUtils from "../../../utils/common/httpUtils";
+import HttpUtils from "../../../utils/api/httpUtils";
 import NotificationUtils from "../../../utils/common/notificationUtils";
 import QueryUtils from "../../../utils/common/queryUtils";
 import React from "react";
