+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/cells/cell/MicroserviceList.js
@@ -17,7 +17,7 @@
 import Constants from "../../../utils/constants";
 import DataTable from "../../common/DataTable";
 import HealthIndicator from "../../common/HealthIndicator";
-import HttpUtils from "../../../utils/common/httpUtils";
+import HttpUtils from "../../../utils/api/httpUtils";
 import {Link} from "react-router-dom";
 import NotificationUtils from "../../../utils/common/notificationUtils";
 import QueryUtils from "../../../utils/common/queryUtils";
