+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/cells/microservice/Details.js
@@ -15,7 +15,7 @@
  */
 
 import HealthIndicator from "../../common/HealthIndicator";
-import HttpUtils from "../../../utils/common/httpUtils";
+import HttpUtils from "../../../utils/api/httpUtils";
 import {Link} from "react-router-dom";
 import NotificationUtils from "../../../utils/common/notificationUtils";
 import QueryUtils from "../../../utils/common/queryUtils";
