+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/tracing/view/SequenceDiagram.css
