+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/tracing/view/SequenceDiagram.js
@@ -16,15 +16,15 @@
 
 import "./SequenceDiagram.css";
 import ChevronRight from "@material-ui/icons/ChevronRight";
-import Constants from "../../../../utils/constants";
+import Constants from "../../../utils/constants";
 import React from "react";
-import Span from "../../../../utils/tracing/span";
-import TracingUtils from "../../../../utils/tracing/tracingUtils";
+import Span from "../../../utils/tracing/span";
+import TracingUtils from "../../../utils/tracing/tracingUtils";
 import Typography from "@material-ui/core/Typography";
 import interact from "interactjs";
 import mermaid from "mermaid";
 import {withStyles} from "@material-ui/core/styles";
-import withColor, {ColorGenerator} from "../../../common/color";
+import withColor, {ColorGenerator} from "../../common/color";
 import * as PropTypes from "prop-types";
 
 const styles = (theme) => ({
