+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/components/tracing/view/index.js
@@ -17,12 +17,12 @@
 import DependencyDiagram from "./DependencyDiagram";
 import ErrorBoundary from "../../common/error/ErrorBoundary";
 import Grey from "@material-ui/core/colors/grey";
-import HttpUtils from "../../../utils/common/httpUtils";
+import HttpUtils from "../../../utils/api/httpUtils";
 import NotFound from "../../common/error/NotFound";
 import NotificationUtils from "../../../utils/common/notificationUtils";
 import Paper from "@material-ui/core/Paper/Paper";
 import React from "react";
-import SequenceDiagram from "./sequenceDiagram/SequenceDiagram";
+import SequenceDiagram from "./SequenceDiagram";
 import Span from "../../../utils/tracing/span";
 import Tab from "@material-ui/core/Tab";
 import Tabs from "@material-ui/core/Tabs";
