+++ system/control-plane/global/components/pom.xml
@@ -26,7 +26,7 @@
 
     <modelVersion>4.0.0</modelVersion>
     <artifactId>org.wso2.vick.global.plane.components</artifactId>
-    <name>WSO2 VICK - Components</name>
+    <name>WSO2 VICK - Global Plane Components</name>
     <packaging>pom</packaging>
 
     <modules>
